+++ clio-server/src/main/scala/org/broadinstitute/clio/server/dataaccess/CachedServerStatusDAO.scala
@@ -1,6 +1,6 @@
 package org.broadinstitute.clio.server.dataaccess
 
-import org.broadinstitute.clio.model.ServerStatusInfo
+import org.broadinstitute.clio.server.model.ServerStatusInfo
 
 import scala.concurrent.Future
 import scala.util.Try
