+++ clio-server/src/main/scala/org/broadinstitute/clio/server/webservice/ExceptionDirectives.scala
@@ -5,7 +5,7 @@ import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server._
 import io.circe.generic.auto._
 import io.circe.syntax._
-import org.broadinstitute.clio.model.ErrorResult
+import org.broadinstitute.clio.server.model.ErrorResult
 
 trait ExceptionDirectives {
   lazy val completeWithInternalErrorJson: Directive0 = {
