+++ clio-server/src/test/scala/org/broadinstitute/clio/server/dataaccess/MockServerStatusDAO.scala
@@ -1,6 +1,6 @@
 package org.broadinstitute.clio.server.dataaccess
 
-import org.broadinstitute.clio.model.ServerStatusInfo
+import org.broadinstitute.clio.server.model.ServerStatusInfo
 
 import scala.concurrent.Future
