+++ clio-server/src/test/scala/org/broadinstitute/clio/server/service/StatusServiceSpec.scala
@@ -2,11 +2,11 @@ package org.broadinstitute.clio.server.service
 
 import org.broadinstitute.clio.server.MockClioApp
 import org.broadinstitute.clio.server.dataaccess.{
-  MockSearchDAO,
   MockHttpServerDAO,
+  MockSearchDAO,
   MockServerStatusDAO
 }
-import org.broadinstitute.clio.model.{StatusInfo, VersionInfo}
+import org.broadinstitute.clio.server.model.{StatusInfo, VersionInfo}
 import org.scalatest.{AsyncFlatSpec, Matchers}
 
 class StatusServiceSpec extends AsyncFlatSpec with Matchers {
