+++ clio-server/src/test/scala/org/broadinstitute/clio/server/webservice/RejectionDirectivesSpec.scala
@@ -4,8 +4,7 @@ import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.testkit.ScalatestRouteTest
 import de.heikoseeberger.akkahttpcirce.FailFastCirceSupport._
 import io.circe.generic.auto._
-import org.broadinstitute.clio.model.RejectionResult
-import org.broadinstitute.clio.server.model.MockResult
+import org.broadinstitute.clio.server.model.{MockResult, RejectionResult}
 import org.scalatest.{FlatSpec, Matchers}
 
 class RejectionDirectivesSpec
